+++ jdaviz/tests/test_line_database.py
@@ -1,0 +1,10 @@
+import numpy as np
+from jdaviz.data.linelists import query_helpers
+
+
+def test_line_database_unique():
+    '''
+    Test that all line names in existing database are unique.
+    '''
+    database = query_helpers.load_db()
+    assert len(database) == len(np.unique(database['line_name']))